simplify CANFIX testing
